requirements.txt
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pkg-resources==0.0.0
